update video carousel and user controller

+++ back-end/src/image/image.service.ts
@@ -6,6 +6,7 @@ import { ProjectEntity } from 'src/project/entity/project-entity';
 import { unlink, access } from 'fs/promises'
 import { existsSync } from 'fs';
 import { ImageEntity } from './entity/image-entity';
+import { ThumbProjectEntity } from './entity/image-project-thumb-entity';
 
 @Injectable()
 export class ImageService {
@@ -13,13 +14,14 @@ export class ImageService {
         @InjectRepository(ImageProjectEntity)
         private imageProjectRepository: Repository<ImageProjectEntity>,
 
+        @InjectRepository(ThumbProjectEntity)
+        private thumbRepository: Repository<ThumbProjectEntity>,
+
         @InjectRepository(ImageEntity)
-        private imageRepositoru: Repository<ImageEntity>,
+        private imageRepository: Repository<ImageEntity>,
 
         @InjectRepository(ProjectEntity)
         private projectRepository: Repository<ProjectEntity>,
-
-        private dataSource: DataSource,
 
 
     ) { }
@@ -31,21 +33,15 @@ export class ImageService {
                 relations: { imageList: true }
             }
         )
-
         if (!project) {
-            await Promise.all(listFile.map(async (file) => {
-                try {
-                    await unlink(file.path);
+            for (const file of listFile) {
+                if (existsSync(file.path)) {
                     await unlink(file.path)
-                    return
-                } catch (error) {
-                    return;
                 }
-            }));
+            }
             throw new NotFoundException('Can not found project');
         }
         const oldImages = project.imageList;
-
         if (oldImages.length > 0) {
             for (const image of oldImages) {
                 if (existsSync(image.path)) {
@@ -54,8 +50,6 @@ export class ImageService {
             }
             await this.imageProjectRepository.remove(oldImages);
         }
-
-
         try {
             const listFileSave: ImageProjectEntity[] = listFile.map((file) => {
                 const fileEntity = this.imageProjectRepository.create({
@@ -68,22 +62,72 @@ export class ImageService {
             await this.imageProjectRepository.save(listFileSave);
             return `Update Images for project ${project.name}`
         } catch (error) {
-            await Promise.all(listFile.map(async (file) => {
-                try {
-                    await access(file.path);
-                    await unlink(file.path)
-                    return
-                } catch (error) {
-                    return;
+            for (const image of oldImages) {
+                if (existsSync(image.path)) {
+                    await unlink(image.path)
                 }
-            }));
+            }
             throw new InternalServerErrorException('Can not upload')
         }
     }
 
-    async getImageById(imageId: number) {
+    async uploadThumbProject(projectId: number, file: Express.Multer.File) {
+        const project = await this.projectRepository.findOne(
+            {
+                where: { id: projectId }, relations: { thumb: true }
+            }
+        )
+        if (!project) {
+            return new NotFoundException('Not Found Project');
+        }
+        const oldThumb = project.thumb;
+        try {
+            if (oldThumb) {
+                await this.thumbRepository.update({ id: oldThumb.id }, { path: file.path });
+                existsSync(oldThumb.path) && await unlink(oldThumb.path);
+            } else {
+                const newThumb = this.thumbRepository.create({
+                    path: file.path,
+                    project
+                });
+                const savedThumb = await this.thumbRepository.save(newThumb);
+                await this.projectRepository.update({ id: projectId }, { thumb: savedThumb })
+            }
+            return `Thumbnail project ${project.name} was upload`;
+        } catch (error) {
+            throw new InternalServerErrorException('Server Error')
+        }
+    }
 
-        return await this.imageRepositoru.findOneByOrFail({ id: imageId })
+    async getImageById(imageId: number) {
+        try {
+            return await this.imageRepository.findOneByOrFail({ id: imageId })
+
+        } catch (error) {
+            throw new NotFoundException('Not found image')
+        }
+
+    }
+
+    async deleteImage(imageId: number, projectId: number) {
+        const image = await this.imageRepository.findOneBy({ id: imageId });
+        if (!image) {
+            throw new NotFoundException('Not found Image')
+        }
+        try {
+            if (image instanceof ThumbProjectEntity) {
+                await this.projectRepository.update({ id: projectId }, { thumb: null });
+                await this.thumbRepository.remove(image);
+            } if (image instanceof ImageProjectEntity) {
+                await this.imageProjectRepository.remove(image)
+            } else {
+                await this.imageRepository.remove(image);
+            }
+            await unlink(image.path);
+            return 'Image was delete';
+        } catch (error) {
+            throw new InternalServerErrorException("Server Errors")
+        }
 
     }
 }
